+++ changelog.rst
@@ -2,6 +2,18 @@
 **CHANGELOG**
 ================
 
+NEW FEATURE: Supercharged RTE
+
+- jsonToHtml function support added
+
+*v0.2.0*
+============
+
+**Date: 02-Sept-2021**
+
+ - Initial release of contentstack utility package
+
+
 ENHANCEMENT, NEW FEATURE, BUG RESOLVE
 
 *v0.1.0*